requirements.txt
flask
webvtt-py
requests
python-dotenv
textblob

